fix(canon): keep the shared editor session correct across disk changes
The reusable editor LSP session made two pre-existing behaviours regress.
tsgo panics with "ScriptKind must be specified when parsing source file" on
a .vue rename, and that crash now takes down the session hover, completion,
and diagnostics share, so only extensions the runtime can parse reach the
native rename and .vue falls back to the import scanner. The session also
pins the files it read while building its program, so watched declaration
and manifest edits plus file create, delete, and rename events now retire
it; .vue content changes still ride their synchronized virtual document.

Splits editor_lsp, its client, and the import-resolution test module to
stay under the 350-line source limit.
