+++ crates/vize_canon/tests/lsp_import_resolution.rs
@@ -3,6 +3,8 @@ use std::path::{Path, PathBuf};
 use vize_canon::{CorsaBridge, CorsaBridgeConfig, LspHover, LspHoverContents, LspMarkedString};
 use vize_carton::ToCompactString;
 
+#[path = "lsp_import_resolution/shared_editor_session.rs"]
+mod shared_editor_session;
 #[path = "lsp_import_resolution/virtual_dependencies.rs"]
 mod virtual_dependencies;
